:truck: [#149] Move service factory to higher level
This way we can add traits for other pre-configured services,
like the gpp search.

Additionally, this configures the gpp-search service for
docker compose, which is to be added later.
